StateChecker: Don't save hash traces (slow and unused)
On every state hash check the StateChecker saved the whole debug_backtrace() and never
removed it again, resulting in slow execution and ever growing memory usage.

Since it wasn't used anywhere just remove it and remove the 'saveTraces' option in
StateCheckerConfig.

This reduces the test time by ~10% and the memory usage by ~70% (on travis-ci).